Fix ipsecPolicies output
Added the additional properties that were missing and corrected the syntax of the output (as the .tf file was malformed previously and threw an error in terraform). 
Note: I have not extensively tested this code, but in my azure resource all of the properties existed. I'm not sure if they will always exist. Regardless, the above amended code results in a valid .tf file that terraform was able to read and no chnages are reported by a terraform plan